Replace primitive type double to a String in Score. This is because of inaccurate representations of doubles in BigDecimals

+++ src/main/java/com/example/demo/adapter/in/service/RatingServiceAdapter.java
@@ -50,7 +50,7 @@ public class RatingServiceAdapter {
                 .build();
 
         PoliticiansRating rating = new PoliticiansRating.Builder()
-                .setRating(Score.of(dtoRequest.getRating().doubleValue()))
+                .setRating(Score.of(dtoRequest.getRating()))
                 .setRater(rater)
                 .build();
